lambda/bedrock-agent-fundasta/requirements.txt
boto3
aws-lambda-powertools
python-dotenv
aioboto3
